refactor(shared): extract @boardsesh/velvet-tokens package
Move the canonical Velvet Send brand palette, surface anchors, and colour
helpers (brandColors, brandColorsDark, materialSurfaces, withAlpha,
blendOpaque) out of the mobile theme into a new pure-TS package shared by
web and mobile. Mobile colors.ts re-exports them so every existing import
path is unchanged; platform-specific resolution (iOS PlatformColor, M3
tonal containers, Android fallbacks, Play-Drawer tint) stays in mobile.

The dev-warning guard uses process.env.NODE_ENV (inlined by Next, populated
by Metro/RN) so the helpers are portable to Next.js. Wires the package into
web (deps + transpilePackages) and mobile deps.

Behaviour-neutral for native: typecheck, 2577 tests, and the Metro bundle
all pass. This is the shared source of truth the web Velvet Send makeover
builds on.
