+++ packages/mobile/package.json
@@ -33,6 +33,7 @@
     "@boardsesh/queue-react": "*",
     "@boardsesh/queue-runtime": "*",
     "@boardsesh/shared-schema": "*",
+    "@boardsesh/velvet-tokens": "*",
     "@expo/vector-icons": "15.1.1",
     "@gorhom/bottom-sheet": "^5.2.14",
     "@react-native-async-storage/async-storage": "^3.1.0",
