+++ packages/shared/velvet-tokens/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@boardsesh/velvet-tokens",
+  "version": "1.0.0",
+  "type": "module",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "exports": {
+    ".": {
+      "types": "./src/index.ts",
+      "import": "./src/index.ts",
+      "default": "./src/index.ts"
+    }
+  },
+  "scripts": {
+    "test": "vp test run",
+    "typecheck": "tsc --noEmit"
+  },
+  "devDependencies": {
+    "@types/node": "^25.3.3",
+    "typescript": "~6.0.3",
+    "vitest": "*"
+  }
+}
